Automation of vpdupdate to provide up-to-date VPD data
This patch is intended for providing an automation for vpdupdate
in case of any changes that happen to device configuration during
run time. During hot plugs, the changes made to the device configuration
are caught by an udev rule and will create a new file in /var/lib/lsvpd/.
When any lsvpd set of tools need vpd information, instead of directly
accessing vpd information from vpd.db file we cross check the modification
times of vpd.db and the file created by udev rule file. If the modification
time of the file created by udev rule is greater than vpd.db this implies
that there is a change to device configuration and vpdupdate is run. To
make this entire process transparent to the tools we do the check in
libvpd, so that the tools don't have to worry about the check.

Changes from V12:
1) Added macros in Makefile.am.

Signed-off-by: Phani Yadav <phayadav@linux.vnet.ibm.com>
Signed-off-by: Suzuki K. Poulose <suzuki@in.ibm.com>