libvpd: New logger macro is introduced
log_msg is a macro which takes priority, logging message as params and logs
messages. It supports variable arguments, so that it can be used with any
number of params.
log_<X> macro is a specific macro, who calls log_msg with priority as a param.
log_<X> macro is able keep macro name as understable with error priority and
can be undefined for any specific priority error.

Signed-off-by: Ankit Kumar <ankit@linux.vnet.ibm.com>
[Minor fix to logger output - Vasant]
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>