libvpd: Corrects data type of variables
In the below snippet the result of comparison between 'string::npos'
(size_t) and 'end' (unsigned int) variable was always a false statement
catched by coverity tool and same is the case with comparison between
'string::npos' and 'z' (unsigned int) variable.

Snippet:
 ...

   if ( string::npos == (end = s1.find("*", beg))) {
	//No more stars - base case
        if (string::npos != (z = s2.find(s1.substr(beg, end), 0))) {
		return true;
	 ...

This patch fixes the issue by changing the data type of 'end' and 'z'
to size_t.

Signed-off-by: Mukesh Ojha <mukesh02@linux.vnet.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>