Imporve error message during mutiple instances of vpdupdate
When multiple instances of vpdupdate runs, only one will get
the access to the database others will get SQLITE_BUSY error
code.

Currently, the error message returned against the SQLITE_BUSY
error is "database is locked" which does not describe the problem
properly.

The database lock situation occurs when an operation could not
continue because of the conflict within the same database connection.
In the case of multiple vpdupdate instances, the database operation is
denied because the database is busy with some other process.

This patch adds a handler for SQLITE_BUSY error to log the appropriate
error message when multiple vpdupdate instances run simultaneously.

Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>
Reviewed-by: Vaibhav Jain <vaibhav@linux.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>