VpdDbEnv: Make the sqlite operations asynchronous
If we use syncrhonous=OFF, the time taken to inserting into the DB
is significantly reduced with safety of data except during OS crashes.
If the pragma is unsupported, the pragma is ignored by the sqlite and
no error/warning displayed with no side effect on the opened handle.

The vpdupdate's most of overhead being the data insertion, this
change makes it return very faster.

Signed-off-by: Shivaprasad G Bhat <sbhat@linux.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>